Eliminated Java warnings
git-svn-id: https://waterfront.svn.sourceforge.net/svnroot/waterfront/trunk/waterfront@160 aeabe084-532a-4f5f-9356-c178316a5679

+++ java/net/sourceforge/waterfront/ide/services/Main.java
@@ -225,7 +225,7 @@ public class Main
         {
           for(int i = 0; i < len; ++i)
           {
-            int asc = (int) s.charAt(i);
+            int asc = s.charAt(i);
             if(asc % 5 == 0)
               doc.setCharacterAttributes(i, 1, sas, true);
             else
